+++ tasks/schedule.sh
@@ -3,20 +3,20 @@ schedule_task() {
   local duration="$2"
   local datetime="$3"
   local current_day="$4"
+  local timezone
 
-  # Calculate the next day based on the current day
-  if [[ "$(uname)" == "Darwin" ]]; then
-    next_day=$(gdate -d "$current_day + 1 day" "+%Y-%m-%d")
-  else
-    next_day=$(date -d "$current_day + 1 day" "+%Y-%m-%d")
-  fi
+  # Get the system's timezone
+  timezone=$(date +%z)
+
+  # Add the timezone to the datetime
+  datetime_with_timezone="${datetime}${timezone:0:3}:${timezone:3:2}"
 
   if [[ -z "$duration" || -z "$datetime" ]]; then
     echo -e "${RED}Error: Missing duration or datetime for task ID $task_id${RESET}"
     return
   fi
 
-  echo -e "${YELLOW}Scheduling task with ID: $task_id (Duration: $duration minutes, Datetime: $datetime)...${RESET}"
+  echo -e "${YELLOW}Scheduling task with ID: $task_id (Duration: $duration minutes, Datetime: $datetime_with_timezone)...${RESET}"
 
   # Get existing labels and task name for the task
   task_data=$(curl -s --request GET \
@@ -25,10 +25,7 @@ schedule_task() {
   task_name=$(echo "$task_data" | jq -r '.content')
   labels=$(echo "$task_data" | jq -r '.labels | join(", ")')
 
-  # Get task duration, handle cases where duration is null or missing
-  task_duration=$(echo "$task_data" | jq -r '.duration.amount // empty')
-
-  # Do not re-schedule if the task name contains "Google-kalenterin tapahtuma"
+  # Skip scheduling if the task name contains "Google-kalenterin tapahtuma"
   if [[ "$task_name" == *"Google-kalenterin tapahtuma"* ]]; then
     echo -e "${YELLOW}Skipping scheduling task: $task_name (ID: $task_id)${RESET}"
     return 0
@@ -39,7 +36,7 @@ schedule_task() {
   due_string=$(echo "$task_data" | jq -r '.due.string')
 
   # Debugging output to check the variables
-  echo "Task name: $(echo "$task_data" | jq -r '.content'), Task ID: $task_id, Duration: $duration, Datetime: $datetime, Recurring: $recurring, Labels: $labels"
+  echo "Task name: $task_name, Task ID: $task_id, Duration: $duration, Datetime: $datetime_with_timezone, Recurring: $recurring, Labels: $labels"
 
   if [ "$recurring" == "true" ]; then
     # Update task's details and keep recurrence
@@ -47,14 +44,14 @@ schedule_task() {
       --url "https://api.todoist.com/rest/v2/tasks/$task_id" \
       --header "Content-Type: application/json" \
       --header "Authorization: Bearer ${TODOIST_API_KEY}" \
-      --data "{\"due_datetime\": \"$datetime\", \"due_string\": \"$due_string\", \"duration\": \"$duration\", \"duration_unit\": \"minute\"}")
+      --data "{\"due_datetime\": \"$datetime_with_timezone\", \"due_string\": \"$due_string\", \"duration\": \"$duration\", \"duration_unit\": \"minute\"}")
   else
-    # Update the task's detauls
+    # Update the task's details
     update_response=$(curl -s --request POST \
       --url "https://api.todoist.com/rest/v2/tasks/$task_id" \
       --header "Content-Type: application/json" \
       --header "Authorization: Bearer ${TODOIST_API_KEY}" \
-      --data "{\"due_datetime\": \"$datetime\", \"duration\": \"$duration\", \"duration_unit\": \"minute\"}")
+      --data "{\"due_datetime\": \"$datetime_with_timezone\", \"duration\": \"$duration\", \"duration_unit\": \"minute\"}")
   fi
 
   # Check if there was an error during the update