+++ tasks/CHANGELOG.md
@@ -3,6 +3,7 @@
 * Never postpone routines
 * Improve prompt
 * Clean up only metadata, leave other parts of the notes intact
+* Fix Linux timezone issue with scheduling
 
 # 1.0.9: 2024-10-14